Fix #561683 – No status whilst copying data DVD to DVD/ISO after 2048MB copied. 	Fix an overflow when we declared the size copied using 2048 instead of 2048ULL
	* src/plugins/cdrkit/burn-readom.c (brasero_readom_read_stderr):
	* src/plugins/cdrtools/burn-readcd.c (brasero_readcd_read_stderr):
	* src/plugins/dvdcss/burn-dvdcss.c
	(brasero_dvdcss_write_image_thread):
	* src/plugins/libburnia/burn-libburn-common.c
	(brasero_libburn_common_status):
	* src/plugins/libburnia/burn-libburn.c:

svn path=/trunk/; revision=1587

+++ src/plugins/cdrkit/burn-readom.c
@@ -68,11 +68,11 @@ brasero_readom_read_stderr (BraseroProcess *process, const gchar *line)
 
 		brasero_job_get_output_type (BRASERO_JOB (readom), &output);
 		if (output.subtype.img_format == BRASERO_IMAGE_FORMAT_BIN)
-			written = sector * 2048;
+			written = (gint64) ((gint64) sector * 2048ULL);
 		else if (output.subtype.img_format == BRASERO_IMAGE_FORMAT_CLONE)
-			written = sector * 2448;
+			written = (gint64) ((gint64) sector * 2448ULL);
 		else
-			written = sector * 2048;
+			written = (gint64) ((gint64) sector * 2048ULL);
 
 		brasero_job_set_written_track (BRASERO_JOB (readom), written);
 
@@ -110,7 +110,6 @@ brasero_readom_read_stderr (BraseroProcess *process, const gchar *line)
 						BRASERO_BURN_ERROR_GENERAL,
 						_("An internal error occured")));
 	}
-
 
 	return BRASERO_BURN_OK;
 }
